+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+* Fix `Konva.RegularPolygon` size calculations.
+
 ## 7.1.8
 
 * Fix incorrect rendering of `TextPath` in some cases. (again)